changed case representation moved World to a separate module
